+++ folly/Makefile.am
@@ -205,7 +205,7 @@ nobase_follyinclude_HEADERS = \
 	io/Cursor-inl.h \
 	io/IOBuf.h \
 	io/IOStreamBuf.h \
-	io/IOStreamBuf.tcc \
+	io/IOStreamBuf-inl.h \
 	io/IOBufQueue.h \
 	io/RecordIO.h \
 	io/RecordIO-inl.h \
